Fixed error on database

+++ my_app/lib/database/accounts.dart
@@ -1,7 +1,7 @@
 class Accounts {
   String user;
   String accNumber; //max 11 char
-  int balance;
+  double balance;
   int cardColor;
 
   Accounts(this.user, this.accNumber, this.balance, this.cardColor);
@@ -12,7 +12,7 @@ List<Accounts> accounts = accountsData
       (item) => Accounts(
           item['user'] as String,
           item['accNumber'] as String,
-          item['balance'] as int,
+          item['balance'] as double,
           item['cardColor'] as int,
     ),).toList();
 
@@ -26,7 +26,7 @@ var accountsData = [
   {
     "user": "Lorem Ipsum",
     "accNumber": "98675346875",
-    "balance": 200,
+    "balance": 200.toDouble(),
     "cardColor": 0xFFFF70A3
   }
 ];
